requirements.txt
streamlit
datascience
deviantart
pandas
pillow
numpy
ipywidgets